refac: consolidate notifications test files
